src/observability/__init__.py
"""
Observability and monitoring for vector database operations with LLMs.

This module provides tools for monitoring vector database operations,
tracking metrics, and ensuring system health in production environments.
"""

from vector_llm_integration.observability.monitoring import (
    setup_monitoring,
    trace_vectordb_operation,
    log_embedding_metrics
)

__all__ = [
    "setup_monitoring",
    "trace_vectordb_operation",
    "log_embedding_metrics"
]
