pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vector_llm_integration"
version = "0.1.0"
description = "Vector Database Integration with Large Language Models"
readme = "README.md"
authors = [
    {name = "AI Research Team"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "chromadb>=0.4.18",
    "tiktoken>=0.5.1",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.2",
    "traceloop-sdk>=0.4.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
]
demo = [
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.0.0",
    "gradio>=3.41.0",
]
vector_dbs = [
    "pinecone-client>=2.2.4",
    "weaviate-client>=3.24.1",
    "pymilvus>=2.3.1",
    "deeplake>=3.8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
